+++ rslib/src/scheduler/queue/builder/mod.rs
@@ -1,6 +1,7 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
+mod burying;
 mod gathering;
 pub(crate) mod intersperser;
 pub(crate) mod sized_chain;
